Fix up regression about changes in MapLookup

+++ app/cred.go
@@ -267,8 +267,8 @@ func GetUrlCred(cred_id, url_id int64) *UrlCred {
 }
 
 func UrlCredNew(in map[string]interface{}) *UrlCred {
-	Cred_id := u.MapLookup(in, "cred_id", -1).(int64)
-	Url_id := u.MapLookup(in, "url_id", -1).(int64)
+	Cred_id := u.MapLookup(in, "cred_id", int64(-1)).(int64)
+	Url_id := u.MapLookup(in, "url_id", int64(-1)).(int64)
 
 	uc := GetUrlCred(Cred_id, Url_id)
 	uc.Note = u.MapLookup(in, "cred_note", "").(string)
@@ -443,7 +443,7 @@ func GetCredentialByID(id int64) *Credential {
 }
 
 func CredentialNew(in map[string]interface{}) *Credential {
-	userID := u.MapLookup(in, "user_id", 0).(int64)
+	userID := u.MapLookup(in, "user_id", int64(0)).(int64)
 	username := u.MapLookup(in, "cred_username", "").(string)
 	password := u.MapLookup(in, "cred_password", "").(string)
 	c := GetCredential(userID, username, password)
